Extracting nano::make_store in to its own files.
