+++ nano/core_test/bootstrap_ascending.cpp
@@ -1,5 +1,6 @@
 #include <nano/lib/stats.hpp>
 #include <nano/node/bootstrap_ascending/service.hpp>
+#include <nano/node/make_store.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
