+++ nano/core_test/confirmation_height.cpp
@@ -1,4 +1,5 @@
 #include <nano/node/election.hpp>
+#include <nano/node/make_store.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
