+++ nano/core_test/node.cpp
@@ -1,5 +1,6 @@
 #include <nano/lib/config.hpp>
 #include <nano/node/election.hpp>
+#include <nano/node/make_store.hpp>
 #include <nano/node/scheduler/component.hpp>
 #include <nano/node/scheduler/manual.hpp>
 #include <nano/node/scheduler/priority.hpp>
