+++ nano/core_test/processor_service.cpp
@@ -1,5 +1,6 @@
 #include <nano/lib/stats.hpp>
 #include <nano/lib/work.hpp>
+#include <nano/node/make_store.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/store.hpp>
 #include <nano/secure/utility.hpp>
