+++ nano/node/CMakeLists.txt
@@ -143,6 +143,8 @@ add_library(
   lmdb/wallet_value.cpp
   logging.hpp
   logging.cpp
+  make_store.hpp
+  make_store.cpp
   network.hpp
   network.cpp
   nodeconfig.hpp
