+++ nano/node/make_store.cpp
@@ -1,0 +1,15 @@
+#include <nano/node/lmdb/lmdb.hpp>
+#include <nano/node/make_store.hpp>
+#include <nano/node/rocksdb/rocksdb.hpp>
+
+#include <boost/filesystem/path.hpp>
+
+std::unique_ptr<nano::store> nano::make_store (nano::logger_mt & logger, boost::filesystem::path const & path, nano::ledger_constants & constants, bool read_only, bool add_db_postfix, nano::rocksdb_config const & rocksdb_config, nano::txn_tracking_config const & txn_tracking_config_a, std::chrono::milliseconds block_processor_batch_max_time_a, nano::lmdb_config const & lmdb_config_a, bool backup_before_upgrade)
+{
+	if (rocksdb_config.enable)
+	{
+		return std::make_unique<nano::rocksdb::store> (logger, add_db_postfix ? path / "rocksdb" : path, constants, rocksdb_config, read_only);
+	}
+
+	return std::make_unique<nano::lmdb::store> (logger, add_db_postfix ? path / "data.ldb" : path, constants, txn_tracking_config_a, block_processor_batch_max_time_a, lmdb_config_a, backup_before_upgrade);
+}
