+++ nano/node/rocksdb/rocksdb.hpp
@@ -3,6 +3,7 @@
 #include <nano/lib/config.hpp>
 #include <nano/lib/logger_mt.hpp>
 #include <nano/lib/numbers.hpp>
+#include <nano/lib/rocksdbconfig.hpp>
 #include <nano/node/rocksdb/account_store.hpp>
 #include <nano/node/rocksdb/block_store.hpp>
 #include <nano/node/rocksdb/confirmation_height_store.hpp>
