+++ nano/qt_test/qt.cpp
@@ -1,3 +1,4 @@
+#include <nano/node/make_store.hpp>
 #include <nano/qt/qt.hpp>
 #include <nano/test_common/network.hpp>
 #include <nano/test_common/system.hpp>
