+++ nano/test_common/ledger.cpp
@@ -1,3 +1,4 @@
+#include <nano/node/make_store.hpp>
 #include <nano/node/node.hpp>
 #include <nano/test_common/ledger.hpp>
